+++ source/hic_hal/stm32/stm32h743ii/extended_features/source/udb_version.c
@@ -14,7 +14,7 @@
 
 
 #if GIT_LOCAL_MODS == 1
-#define GIT_LOCAL_MODS_STR ""
+#define GIT_LOCAL_MODS_STR "_modified"
 #else
 #define GIT_LOCAL_MODS_STR ""
 #endif //GIT_LOCAL_MODS
